typescript standard-entity-definition
